+++ tests/show_ip_route_common.py
@@ -930,4 +930,5 @@ Routing entry for 0.0.0.0/0
   Last update 01:01:51 ago
   * 10.0.0.7, via PortChannel106
   * 10.0.0.1, via PortChannel102
+
 """
